int to double

+++ optimal_control_casadi/eocarBiorbd.cpp
@@ -34,8 +34,8 @@ int main(int argc, char *argv[]){
     std::vector<std::pair<void (*)(const ProblemSize&,
                          const std::vector<casadi::MX>&,
                          const std::vector<casadi::MX>&,
-                         int,
-                         casadi::MX&), int>> objectiveFunctions;
+                         double,
+                         casadi::MX&), double>> objectiveFunctions;
     objectiveFunctions.push_back(std::make_pair(minimizeTorqueControls, 1000));
     objectiveFunctions.push_back(std::make_pair(minimizeMuscleControls, 1));
     objectiveFunctions.push_back(std::make_pair(minimizeStates, 1.0/1000));
